requirements.txt
cryptography>=41.0.1
rich>=13.3.5
click>=8.1.3
python-dotenv>=1.0.0
pyyaml>=6.0.1
regex>=2023.5.5
sqlalchemy>=2.0.15
passlib>=1.7.4
typer>=0.9.0
pydantic>=2.4.2
python-slugify>=8.0.1 